Demo code structure refactored
