+++ demo/src/components/book-details/index.js
@@ -1,0 +1,3 @@
+import { BookDetails } from "./BookDetails";
+
+export default BookDetails;
