+++ demo/src/components/books-list/index.js
@@ -1,0 +1,3 @@
+import { BooksList } from "./BooksList";
+
+export default BooksList;
