+++ demo/src/components/books-search-bar/index.js
@@ -1,0 +1,3 @@
+import { BooksSearchBar } from "./BooksSearchBar";
+
+export default BooksSearchBar;
