+++ demo/src/components/books-search/index.js
@@ -1,0 +1,3 @@
+import { BooksSearch } from "./BooksSearch";
+
+export default BooksSearch;
