+++ demo/src/components/create-book/index.js
@@ -1,0 +1,3 @@
+import { CreateBook } from "./CreateBook";
+
+export default CreateBook;
