+++ demo/src/components/delete-book/DeleteBook.js
