+++ demo/src/components/delete-book/index.js
@@ -1,0 +1,3 @@
+import { DeleteBook } from "./DeleteBook";
+
+export default DeleteBook;
