+++ demo/src/components/error/index.js
@@ -1,0 +1,3 @@
+import { ErrorComponent } from "./Error";
+
+export default ErrorComponent;
