+++ demo/src/components/not-found/index.js
@@ -1,0 +1,3 @@
+import { NotFound } from "./NotFound";
+
+export default NotFound;
