+++ demo/src/components/update-book/index.js
@@ -1,0 +1,3 @@
+import { UpdateBook } from "./UpdateBook";
+
+export default UpdateBook;
