+++ demo/src/data/books/actions.js
@@ -1,18 +1,5 @@
-import { Api } from "@xbyorange/mercury-api";
-
-import { booksCollection } from "./booksCollection";
-
-// BOOKS MODEL
-
-export const booksModels = new Api(`/books/:id`);
-
-booksModels.addCustomQuery({
-  byId: id => ({
-    urlParams: {
-      id
-    }
-  })
-});
+import { booksModels } from "./origins/booksModels";
+import { booksCollection } from "./origins/booksCollection";
 
 // RELATIONSHIP BETWEEN MODEL AND COLLECTION
