+++ demo/src/data/books/selectors/booksCollection.js
@@ -1,26 +1,6 @@
 import { sortBy } from "lodash";
 import { Selector } from "@xbyorange/mercury";
-import { Api } from "@xbyorange/mercury-api";
-
-// BOOKS COLLECTION
-
-export const booksCollection = new Api(`/books`, {
-  defaultValue: []
-});
-
-const titleContainingFilter = titlePortion => {
-  if (titlePortion && titlePortion.length) {
-    return {
-      queryString: {
-        title_containing: titlePortion
-      }
-    };
-  }
-};
-
-booksCollection.addCustomQuery({
-  titleContaining: titleContainingFilter
-});
+import { booksCollection, titleContainingFilter } from "../origins/booksCollection";
 
 // SELECTORS
