+++ demo/src/index.js
@@ -3,6 +3,6 @@ import ReactDOM from "react-dom";
 
 import "./index.css";
 
-import { App } from "./app/App.js";
+import { App } from "./app/index.js";
 
 ReactDOM.render(<App />, document.getElementById("root"));
