+++ demo/src/modules/books/controllers/CreateBook.js
@@ -1,8 +1,8 @@
 import { connect } from "@xbyorange/react-mercury";
 
-import { booksCollection } from "../../../data-sources/booksCollection";
+import { booksCollection } from "../../../data/books";
 
-import { CreateBook as CreateBookComponent } from "../../../components/books/CreateBook";
+import { default as CreateBookComponent } from "../../../components/create-book";
 
 export const mapDataSourceToProps = () => ({
   createBook: booksCollection.create
