+++ demo/src/modules/books/controllers/DeleteBook.js
@@ -1,8 +1,8 @@
 import { connect } from "@xbyorange/react-mercury";
 
-import { booksModels } from "../../../data-sources/booksModels";
+import { booksModels } from "../../../data/books";
 
-import { DeleteBook as DeleteBookComponent } from "../../../components/books/DeleteBook";
+import { default as DeleteBookComponent } from "../../../components/delete-book";
 
 export const mapDataSourceToProps = ({ id }) => {
   const dataSource = booksModels.byId(id).delete;
