+++ demo/src/modules/books/index.js
@@ -1,0 +1,9 @@
+import { MainRouter } from "./routers/main";
+export default MainRouter;
+
+//No validos por el plugin:
+
+//export { MainRouter } from "./routers/main";
+
+/*import { MainRouter } from "./routers/main";
+export { MainRouter };*/
